refactor: move function tool manager
