+++ core/workflow/workflow_context.py
@@ -8,7 +8,7 @@ from ..adapter import AdapterManager
 from core.chat.session_manager import SessionManager
 from core.config import KiraConfig
 from core.event_bus import EventBus
-from core.llm_client import FuncToolManager
+from core.agent.func_tool_manager import FuncToolManager
 from core.persona import PersonaManager
 from core.utils.path_utils import get_data_path
